\name{fips2010}
\alias{fips2010}
\docType{data}
\title{
FIPS for all levels in the year 2010
}
\description{
FIPS codes for all geographic levels of the US Census 2010. For use with API function.
}
\usage{data("fips2010")}
\source{
\url{https://www.census.gov/geo/reference/codes/cou.html}
}
\examples{
data(fips2010)
}
\keyword{datasets}
