\name{states.names.cap}
\alias{states.names.cap}
\docType{data}
\title{
States Names Capitalized
}
\description{
A list of all the states available in \code{UScensus2010}
}
\usage{data(states.names.cap)}
\format{
  The format is:
 chr [1:49] "Alabama" "Arizona" "Arkansas" "California" "Colorado" ...
}
\details{
For use in the functions of \code{UScensus2010}.
}
%\source{}
%\references{}
\examples{
data(states.names.cap)
}
\keyword{datasets}
