\name{states.names}
\alias{states.names}
\docType{data}
\title{
States Names
}
\description{
A list of all the states available in \code{UScensus2010}
}
\usage{data(states.names)}
\format{
  The format is:
 chr [1:49] "alabama" "arizona" "arkansas" "california" "colorado" ...
}
\details{
For use in the functions of \code{UScensus2010}.
}
%\references{}

\examples{
data(states.names)

}
\keyword{datasets}
