\name{states}
\alias{states}
\docType{data}
\title{
States Names
}
\description{
A list of all the states available in \code{UScensus2010}
}
\usage{data(states.names)}
\format{
  The format is: SpatialPolygonDataFrames
}
\details{
For use in the functions of \code{UScensus2010}.
}
%\references{}

\examples{
data(states)

}
\keyword{datasets}
