\name{UScensus2010-package}
\alias{UScensus2010-package}
\alias{UScensus2010}
\docType{package}
\title{
Helper functions for the UScensus2010-suite of packages
}
\description{
This package helps the user handle the spatial and demographic data available in \code{UScensus2010tract}, \code{UScensus2010cdp}, \code{UScensus2010county}, \code{UScensus2010blkgrp}, and \code{UScensus2010blk}
}
\details{
\tabular{ll}{
Package: \tab UScensus2010\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-18-11\cr
License: \tab GPL Version 2 or newer\cr
LazyLoad: \tab yes\cr
}

}
\author{
Zack W Almquist<almquist@uci.edu> \cr
Maintainer: Zack W Almquist <almquist@uci.edu>
}
\note{
If you use this package and/or software manual in your work, a citation would be appreciated.  References to the current versions are:\cr
\cr
Preferred Citation for the Package: \cr \cr
Zack W. Almquist (2010). US Census Spatial and Demographic Data in R: The UScensus2010 Suite of Packages. Journal of Statistical Software, 37(6), 1-31. \url{http://www.jstatsoft.org/v37/i06/}. \cr \cr

Package Citation: \cr \cr
Zack W Almquist (2010). UScensus2010: US Census 2010 Suite of R Packages. R package version 1.00. \cr

Please also cite the original data source and the \code{\link[sp]{sp}} and \code{maptools}. 
\cr \cr

Census 2010 Summary File 1 [name of state1 or United States]/prepared by the U.S. Census 
Bureau, 2010. 
\cr  \cr

Bivand RS, Pebesma EJ, Gomez-Rubio V (2008). Applied Spatial Data Analysis with R. Springer, New York, NY.


}

%\references{}
\keyword{usUScensus2010}
%\seealso{}
%\examples{}
